+++ client/src/api/Comment/index.js
@@ -1,2 +1,3 @@
+export * from './delete';
 export * from './post';
 export * from './put';
