--- client/src/api/put.js
@@ -1,22 +1,0 @@
-import axiosInstance from '../utils/useAxiosPrivate';
-
-export async function putAnswer(data) {
-  try {
-    await axiosInstance.put(`/answers/${data.answerId}`, {
-      content: data.content,
-    });
-    alert('답변이 수정되었습니다.');
-  } catch (err) {
-    console.log(err);
-  }
-}
-export async function putComment(data) {
-  try {
-    await axiosInstance.put(`/comments/${data.commentId}`, {
-      content: data.content,
-    });
-    alert('댓글이 수정되었습니다.');
-  } catch (err) {
-    console.log(err);
-  }
-}
