+++ client/src/components/AnswerEditModal.js
@@ -5,9 +5,8 @@ import Typography from '@mui/material/Typography';
 import Modal from '@mui/material/Modal';
 import { IconButton } from '@mui/material';
 import { useForm } from 'react-hook-form';
-import { postAnswer } from '../api/post';
+import { postAnswer, putAnswer } from '../api/Answer';
 import { useNavigate, useParams } from 'react-router-dom';
-import { putAnswer } from '../api/put';
 import { useEffect, useState, useMemo } from 'react';
 import ClearIcon from '@mui/icons-material/Clear';
 import { answerStore } from '../store/store';
