+++ client/src/pages/PostPage/Answer.js
@@ -14,8 +14,7 @@ import { useParams } from 'react-router-dom';
 import AnswerEditModal from '../../components/AnswerEditModal';
 import styled from 'styled-components';
 import { useForm } from 'react-hook-form';
-import { postComment } from '../../api/post';
-import { deleteComment } from '../../api/delete';
+import { postComment, deleteComment } from '../../api/Comment';
 import CommentEditForm from '../../components/CommentEditForm';
 
 export default function Answer(props) {
