+++ client/src/utils/axiosRequest.js
@@ -9,72 +9,13 @@ import {
   setUserId,
   removeUserId,
 } from './cookies';
-import axiosInstance from './useAxiosPrivate';
-
-export async function postSignUp(data) {
-  try {
-    const res = await axiosInstance.post('/users', data);
-    alert('회원가입이 완료되었습니다!');
-  } catch (err) {
-    alert(err.response.data.message);
-  }
-}
-export async function postSignIn(data) {
-  try {
-    const res = await axiosInstance.post('/login', data);
-    setAccessToken(res.data.accessToken);
-    setRefreshToken(res.data.refreshToken);
-    setUserId(res.data.userId);
-    alert('로그인이 성공했습니다');
-  } catch (err) {
-    alert(err.response.data.message);
-  }
-}
+import { postLogout } from '../api/User';
+import axiosInstance from './axiosInstance';
 
 export async function postQuestion(data) {
   try {
     const res = await axiosInstance.post('/questions', data);
     alert('질문이 등록되었습니다.');
-  } catch (err) {
-    alert(err.response.data.message);
-  }
-}
-
-export async function getAccessWithRefresh() {
-  try {
-    const instance = axiosInstance.create({
-      headers: { Refresh: `${getRefreshToken()}` },
-    });
-    const res = await instance.get('/refresh');
-    await removeAccessToken();
-    setAccessToken(res.headers.authorization);
-    return res.headers.authorization;
-  } catch (err) {
-    console.log(err);
-    if (
-      (err?.response?.status === 401 &&
-        err?.response?.data?.message === 'REFRESH TOKEN EXPIRED') ||
-      (err?.response?.status === 400 &&
-        err?.response?.data?.message === 'INVALID REFRESH TOKEN')
-    ) {
-      return postLogout();
-    }
-  }
-}
-
-export async function postLogout() {
-  try {
-    const res = await axiosInstance.post(
-      '/auths/logout',
-      {},
-      {
-        headers: { Authorization: `${getRefreshToken()}` },
-      }
-    );
-    removeAccessToken();
-    removeRefreshToken();
-    removeUserId();
-    alert('로그아웃 되었습니다');
   } catch (err) {
     alert(err.response.data.message);
   }
@@ -104,25 +45,5 @@ export async function updateAnswerVotes(answerId, votes) {
   } catch (err) {
     alert(err.response.data.message);
     console.log(err.response);
-  }
-}
-
-export async function modifyUser(nickname, profile) {
-  try {
-    const res = await axiosInstance.put('/users/modify', {
-      profile: profile,
-      nickname: nickname,
-    });
-  } catch (err) {
-    console.log(err);
-  }
-}
-
-export async function modifyPassword(data) {
-  try {
-    const res = await axiosInstance.put('/users/password', data);
-    alert('비밀번호 변경이 완료되었습니다!');
-  } catch (err) {
-    alert('기존의 비밀번호를 확인해주세요');
   }
 }
